requirements.txt
openai>=1.0.0
mcp>=1.0.0
nltk>=3.8.0
python-dotenv>=1.0.0
pydantic>=2.0.0 